pyproject.toml
[project]
name = "reef-server"
version = "1.0.0"
description = "FastAPI server for Reef iOS app - proxies Gemini API calls"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "lib"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
